requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
attrs==25.1.0
audioop-lts==0.2.1
discord.py[voice]==2.5.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
python-dotenv==1.0.1
setuptools==75.8.0
typing_extensions==4.12.2
wheel==0.45.1
yarl==1.18.3
yt_dlp==2025.2.19
